rename hooks files to ts file
